requirements.txt
streamlit
tensorflow
numpy
pillow
streamlit-webrtc
opencv-python-headless